Return a minimum allocation when size is 0
The original version just did the easier thing allowed by returning the
null pointer for a zero size. This was POSIX acceptable, but glibc
explicitly returns a unique pointer that is valid to give back to free.

GNU grep assumes glibc behavior: if malloc returns the null pointer, it
dies with OOM. Other GNU utilities are bound to depend on this behavior,
so, you know, whatever.

It would be instructive to check the behavior of other allocators and
see what the "canonical" behavior is.